requirements.txt
aniso8601==8.0.0
click==7.1.2
cssselect2==0.3.0
Flask==1.1.2
Flask-RESTful==0.3.8
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.5.0
MarkupSafe==1.1.1
Pillow==7.1.2
PyPDF2==1.26.0
pytz==2020.1
reportlab==3.5.42
six==1.14.0
svglib==1.0.0
tinycss2==1.0.2
webencodings==0.5.1
Werkzeug==1.0.1
pandas==1.0.3
xlwt==1.3.0
xlrd==1.2.0
psycopg2-binary==2.8.5
pydrive==1.3.1